added stereoenhance and mopidy source (not ready yet)

+++ src/server/inputmux.py
@@ -21,9 +21,10 @@ class InputMux(util.Threadbase):
     Hosts the sources (currently sourcefifo, sourcetcp and sourcespotifyd), listens to their events
     and makes sure that the events passed on are sane.
     """
-    def __init__(self):
+    def __init__(self, source_config):
         super(InputMux, self).__init__(name='inputmux  ')
         self.queue = queue.Queue()
+        self.source_event_lock = threading.Lock()
 
         self.log_first_audio = LOG_FIRST_AUDIO_COUNT
         self.now_playing = None
@@ -32,12 +33,17 @@ class InputMux(util.Threadbase):
 
         self.sourcefifo = sourcefifo.SourceFifo()
         self.sourcefifo.connect("event", self.input_event)
-        self.sourcetcp = sourcetcp.SourceTCP(4665)  # fixit
+        self.sourcetcp = sourcetcp.SourceTCP(source_config['gstreamer_port'])
         self.sourcetcp.connect("event", self.input_event)
         self.sourcespotifyd = sourcespotifyd.SourceSpotifyd()
         self.sourcespotifyd.connect("event", self.input_event)
-        self.source_mopidy = sourcemopidy.SourceMopidy('192.168.1.102', 4666) # fixit
-        self.source_mopidy.connect("event", self.input_event)
+        if source_config['mopidy_ws_enabled'] == 'on':
+            self.source_mopidy = sourcemopidy.SourceMopidy(source_config['mopidy_ws_address'],
+                                                           source_config['mopidy_ws_port'],
+                                                           source_config['mopidy_gst_port'])
+            self.source_mopidy.connect("event", self.input_event)
+        else:
+            self.source_mopidy = None
 
         threading.Thread(target=self.gst_mainloop_thread).start()
         self.start()
@@ -47,7 +53,8 @@ class InputMux(util.Threadbase):
         self.sourcefifo.terminate()
         self.sourcetcp.terminate()
         self.sourcespotifyd.terminate()
-        self.source_mopidy.terminate()
+        if self.source_mopidy:
+            self.source_mopidy.terminate()
         self.mainloop.quit()
 
     @staticmethod
@@ -72,36 +79,37 @@ class InputMux(util.Threadbase):
         Once playing it buffers audio to sane package sizes. Outdata (messages and audio) are
         placed in a queue for others to pick up (currently picked by the server itself)
         """
-        source, key, value = skv
+        with self.source_event_lock:
+            source, key, value = skv
 
-        if key == 'codec':
-            if self.now_playing:
-                log.info('inputmux ditching current source "%s", new codec=%s' % (self.now_playing, value))
-                self.stop_playing()
-            self.now_playing = source
-            self.log_first_audio = LOG_FIRST_AUDIO_COUNT
-            log.info('inputmux starts playing source "%s"' % source)
-            self.audio_buffer = bytearray()
+            if key == 'codec':
+                if self.now_playing:
+                    log.info('inputmux ditching current source "%s", new codec=%s' % (self.now_playing, value))
+                    self.stop_playing()
+                self.now_playing = source
+                self.log_first_audio = LOG_FIRST_AUDIO_COUNT
+                log.info('inputmux starts playing source "%s"' % source)
+                self.audio_buffer = bytearray()
 
-        if not self.now_playing or self.now_playing != source:
-            return
-
-        if key == 'audio':
-            if self.log_first_audio:
-                self.log_first_audio -= 1
-                log.debug('audio %s bytes' % len(value))
-            self.timeout_counter = 40
-            self.audio_buffer += value
-            if len(self.audio_buffer) < 3000:
+            if not self.now_playing or self.now_playing != source:
                 return
-            value = self.audio_buffer
-            self.audio_buffer = bytearray()
 
-        self.queue.put_nowait([key, value])
+            if key == 'audio':
+                if self.log_first_audio:
+                    self.log_first_audio -= 1
+                    log.debug('audio %s bytes' % len(value))
+                self.timeout_counter = 40
+                self.audio_buffer += value
+                if len(self.audio_buffer) < 3000:
+                    return
+                value = self.audio_buffer
+                self.audio_buffer = bytearray()
 
-        if key == 'state' and value == 'stop':
-            log.info('inputmux received stop from source "%s"' % source)
-            self.stop_playing()
+            self.queue.put_nowait([key, value])
+
+            if key == 'state' and value == 'stop':
+                log.info('inputmux received stop from source "%s"' % source)
+                self.stop_playing()
 
     def run(self):
         while not self.terminated:
